requirement.txt
certifi==2019.11.28
chardet==3.0.4
EasyProcess==0.2.8
idna==2.8
Pillow==6.2.1
pytesseract==0.3.1
PyVirtualDisplay==0.2.5
requests==2.22.0
selenium==3.141.0
urllib3==1.25.7
